converted service generator to a separate file
